Addition of input + planning
The dates for the planning aren't added yet.
Addition of the compiler.
addition of the first modules to read and write the inputs.

All necessary datas can now be read with the necessary conditions set to them.
The necessities take care of the possibility to have addition of X different compounds.

I need to sleep.