Addition of dimer calculation and searcher+bug correction
Addition of sigma_epsilon_dimer() that compute all the possible dimers of sigma and epsilon and allow a smoother calculation for a higher number of species.
Addition of pick_dimers_data() that will find inside the matrix of dimers the exact position of each dimers without using indexes.
This function is believed to me more efficient at high number of species in opposition to putting the calculation of each dimers in the loop.

Bug correction, namely:
- geometric_mean()
- random_select()

Test for all the functions created by Aledde, and edits to let them use global variables.