V1 calculation of the energy
Here is the correction of the energy calculation to avoid the computation of the energy of a compound with itself.

The code however feels odd because each time the old Energy is equal to zero, almost all of the time. (No infinite test were done, so, it's only a supposition).
A V2 will be proposed trying to solve this problem