V3 - LOG ADDITION + ADDITION OF THEO_PROGRAM FOR N COMPOUNDS
Addition of the log system with input and output working.

Addition of a function to output the energies.
Addition of the functions to compute until N particles.

The function need to be tested for SINGLE COMPOUND LJ fluids and BI-LJ fluids.

Redo of the out_energy.txt to make it lighter...

And...  I still need to do the Lab report...
Well... Good luck to the person that will need to the g(r).

I will add the function that will output the g(r) when it will be necessary.

But for now, the energies are in theory corrected for N compounds!